requirements.txt
fastapi
uvicorn
langchain
openai
pypdf
tiktoken
pinecone-client
python-dotenv
requests
